test: :test_tube: :recycle: move tests of float functionality
