+++ tests/test_triplets_that_sum.py
@@ -1,4 +1,3 @@
-from math import e, pi, sqrt
 import pytest
 from tuplesumfilter import triplets_that_sum_to
 
@@ -30,13 +29,6 @@ def test_triplets_worked_example(worked_example_nums):
 
 def test_triplets_with_finite_input_but_no_match(worked_example_nums):
     assert triplets_that_sum_to(worked_example_nums, 0) == []
-
-
-def test_triplets_works_approx_with_floats():
-    nums = [pi, sqrt(2), 17.45, 1e-10, e]
-    target = pi + e + sqrt(2)
-    expected = [(pytest.approx(pi), pytest.approx(sqrt(2)), pytest.approx(e))]
-    assert triplets_that_sum_to(nums, target) == expected
 
 
 def test_triplets_when_multiple_matches():
